refactor: VacancyDetail components refactoring
