+++ src/features/vacancy-detail/ui/components/VacancyMainContent/VacancyMainContent.types.ts
@@ -1,0 +1,5 @@
+import type { Vacancy } from '@/entities/vacancy';
+
+export interface VacancyMainContentProps {
+  vacancy: Vacancy;
+}
