+++ src/features/vacancy-detail/ui/components/VacancySummaryCard/VacancySummaryCard.types.ts
@@ -1,0 +1,5 @@
+import type { Vacancy } from '@/entities/vacancy';
+
+export interface VacancySummaryCardProps {
+  vacancy: Vacancy;
+}
