+++ src/features/vacancy-detail/ui/components/index.ts
@@ -1,0 +1,4 @@
+export { VacancyMainContent } from './VacancyMainContent/VacancyMainContent';
+export type { VacancyMainContentProps } from './VacancyMainContent/VacancyMainContent.types';
+export { VacancySummaryCard } from './VacancySummaryCard/VacancySummaryCard';
+export type { VacancySummaryCardProps } from './VacancySummaryCard/VacancySummaryCard.types';
